Create datatime helper function
